we are no longer shipping MS.IO.Redist, there is no need for File to be partial
